TIP-882: move ref data checkers
